requirements.txt
# Minimal working requirements
python-dotenv==1.0.0
discord-webhook==1.3.0
aiohttp==3.9.1
pandas==2.1.4
SQLAlchemy==2.0.23
python-json-logger==2.0.7
apscheduler==3.10.4
python-binance==1.0.19
websocket-client==1.6.4
